feat: include option to skip duplicated assets
This prevents deleting assets with the same name

+++ index.js
@@ -154,25 +154,27 @@ PublishRelease.prototype.publish = function publish () {
             if (res.statusCode === 422 && bodyJson.errors && bodyJson.errors[0].code === 'already_exists') {
               self.emit('duplicated-asset', fileName)
 
-              async.eachSeries(obj.createRelease.assets, function (el) {
-                if (fileName === el.name) {
-                  const deleteAssetUri = obj.createRelease.url.split('/').slice(0, -1).join('/') + '/assets/' + el.id
+              if (!opts.skipDuplicatedAssets) {
+                async.eachSeries(obj.createRelease.assets, function (el) {
+                  if (fileName === el.name) {
+                    const deleteAssetUri = obj.createRelease.url.split('/').slice(0, -1).join('/') + '/assets/' + el.id
 
-                  request({
-                    method: 'DELETE',
-                    uri: deleteAssetUri,
-                    headers: {
-                      'Authorization': 'token ' + opts.token,
-                      'User-Agent': 'publish-release ' + pkg.version + ' (https://github.com/remixz/publish-release)'
-                    }
-                  }, function (err, res, body) {
-                    if (err) return callback(err)
+                    request({
+                      method: 'DELETE',
+                      uri: deleteAssetUri,
+                      headers: {
+                        'Authorization': 'token ' + opts.token,
+                        'User-Agent': 'publish-release ' + pkg.version + ' (https://github.com/remixz/publish-release)'
+                      }
+                    }, function (err, res, body) {
+                      if (err) return callback(err)
 
-                    self.emit('duplicated-asset-deleted', fileName)
-                    requestUploadAsset()
-                  })
-                }
-              })
+                      self.emit('duplicated-asset-deleted', fileName)
+                      requestUploadAsset()
+                    })
+                  }
+                })
+              }
             } else {
               self.emit('uploaded-asset', fileName)
               callback()
